requirements.txt
fastapi==0.110.0
sqlalchemy==2.0.29
psycopg2==2.9.9
pydantic==2.6.4
python-multipart==0.0.9
PyJWT==2.8.0
uvicorn==0.29.0
typing_extensions==4.10.0
starlette==0.36.3
bcrypt==4.1.2
httpx==0.27.0